Added all classes containing lat/long functions to lat_long_transform_primitive; added all classes containing date and/or time functions to date_time_transform_primitive
