+++ featuretools/primitives/standard/api.py
@@ -2,6 +2,7 @@
 from .aggregation_primitives import *
 from .binary_transform import *
 from .cum_transform_feature import *
-from .lat_long_primitive import *
+from .date_time_transform_primitive import *
+from .lat_long_transform_primitive import *
 from .rolling_transform_primitive import *
 from .transform_primitive import *
