+++ featuretools/primitives/standard/transform_primitive.py
@@ -1,29 +1,16 @@
-import warnings
-
 import numpy as np
 import pandas as pd
 from woodwork.column_schema import ColumnSchema
 from woodwork.logical_types import (
     URL,
-    AgeFractional,
     Boolean,
     BooleanNullable,
     Categorical,
-    Datetime,
-    Double,
     EmailAddress,
-    LatLong,
-    NaturalLanguage,
-    Ordinal
+    NaturalLanguage
 )
 
 from featuretools.primitives.base import TransformPrimitive
-from featuretools.primitives.utils import (
-    HolidayUtil,
-    _deconstrct_latlongs,
-    _haversine_calculate
-)
-from featuretools.utils import convert_time_units
 from featuretools.utils.common_tld_utils import COMMON_TLDS
 from featuretools.utils.gen_utils import Library
 
@@ -64,274 +51,6 @@ class Absolute(TransformPrimitive):
 
     def get_function(self):
         return np.absolute
-
-
-class TimeSincePrevious(TransformPrimitive):
-    """Compute the time since the previous entry in a list.
-
-    Args:
-        unit (str): Defines the unit of time to count from.
-            Defaults to Seconds. Acceptable values:
-            years, months, days, hours, minutes, seconds, milliseconds, nanoseconds
-
-    Description:
-        Given a list of datetimes, compute the time in seconds elapsed since
-        the previous item in the list. The result for the first item in the
-        list will always be `NaN`.
-
-    Examples:
-        >>> from datetime import datetime
-        >>> time_since_previous = TimeSincePrevious()
-        >>> dates = [datetime(2019, 3, 1, 0, 0, 0),
-        ...          datetime(2019, 3, 1, 0, 2, 0),
-        ...          datetime(2019, 3, 1, 0, 3, 0),
-        ...          datetime(2019, 3, 1, 0, 2, 30),
-        ...          datetime(2019, 3, 1, 0, 10, 0)]
-        >>> time_since_previous(dates).tolist()
-        [nan, 120.0, 60.0, -30.0, 450.0]
-    """
-    name = "time_since_previous"
-    input_types = [ColumnSchema(logical_type=Datetime, semantic_tags={'time_index'})]
-    return_type = ColumnSchema(semantic_tags={'numeric'})
-    description_template = "the time since the previous instance of {}"
-
-    def __init__(self, unit="seconds"):
-        self.unit = unit.lower()
-
-    def get_function(self):
-        def pd_diff(values):
-            return convert_time_units(values.diff().apply(lambda x: x.total_seconds()), self.unit)
-        return pd_diff
-
-
-class Day(TransformPrimitive):
-    """Determines the day of the month from a datetime.
-
-    Examples:
-        >>> from datetime import datetime
-        >>> dates = [datetime(2019, 3, 1),
-        ...          datetime(2019, 3, 3),
-        ...          datetime(2019, 3, 31)]
-        >>> day = Day()
-        >>> day(dates).tolist()
-        [1, 3, 31]
-    """
-    name = "day"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(logical_type=Ordinal(order=list(range(1, 32))), semantic_tags={'category'})
-    compatibility = [Library.PANDAS, Library.DASK, Library.KOALAS]
-    description_template = "the day of the month of {}"
-
-    def get_function(self):
-        def day(vals):
-            return vals.dt.day
-        return day
-
-
-class Hour(TransformPrimitive):
-    """Determines the hour value of a datetime.
-
-    Examples:
-        >>> from datetime import datetime
-        >>> dates = [datetime(2019, 3, 1),
-        ...          datetime(2019, 3, 3, 11, 10, 50),
-        ...          datetime(2019, 3, 31, 19, 45, 15)]
-        >>> hour = Hour()
-        >>> hour(dates).tolist()
-        [0, 11, 19]
-    """
-    name = "hour"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(logical_type=Ordinal(order=list(range(24))), semantic_tags={'category'})
-    compatibility = [Library.PANDAS, Library.DASK, Library.KOALAS]
-    description_template = 'the hour value of {}'
-
-    def get_function(self):
-        def hour(vals):
-            return vals.dt.hour
-        return hour
-
-
-class Second(TransformPrimitive):
-    """Determines the seconds value of a datetime.
-
-    Examples:
-        >>> from datetime import datetime
-        >>> dates = [datetime(2019, 3, 1),
-        ...          datetime(2019, 3, 3, 11, 10, 50),
-        ...          datetime(2019, 3, 31, 19, 45, 15)]
-        >>> second = Second()
-        >>> second(dates).tolist()
-        [0, 50, 15]
-    """
-    name = "second"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(logical_type=Ordinal(order=list(range(60))), semantic_tags={'category'})
-    compatibility = [Library.PANDAS, Library.DASK, Library.KOALAS]
-    description_template = "the seconds value of {}"
-
-    def get_function(self):
-        def second(vals):
-            return vals.dt.second
-        return second
-
-
-class Minute(TransformPrimitive):
-    """Determines the minutes value of a datetime.
-
-    Examples:
-        >>> from datetime import datetime
-        >>> dates = [datetime(2019, 3, 1),
-        ...          datetime(2019, 3, 3, 11, 10, 50),
-        ...          datetime(2019, 3, 31, 19, 45, 15)]
-        >>> minute = Minute()
-        >>> minute(dates).tolist()
-        [0, 10, 45]
-    """
-    name = "minute"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(logical_type=Ordinal(order=list(range(60))), semantic_tags={'category'})
-    compatibility = [Library.PANDAS, Library.DASK, Library.KOALAS]
-    description_template = "the minutes value of {}"
-
-    def get_function(self):
-        def minute(vals):
-            return vals.dt.minute
-        return minute
-
-
-class Week(TransformPrimitive):
-    """Determines the week of the year from a datetime.
-
-    Description:
-        Returns the week of the year from a datetime value. The first week
-        of the year starts on January 1, and week numbers increment each
-        Monday.
-
-    Examples:
-        >>> from datetime import datetime
-        >>> dates = [datetime(2019, 1, 3),
-        ...          datetime(2019, 6, 17, 11, 10, 50),
-        ...          datetime(2019, 11, 30, 19, 45, 15)]
-        >>> week = Week()
-        >>> week(dates).tolist()
-        [1, 25, 48]
-        """
-    name = "week"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(logical_type=Ordinal(order=list(range(1, 54))), semantic_tags={'category'})
-    compatibility = [Library.PANDAS, Library.DASK, Library.KOALAS]
-    description_template = "the week of the year of {}"
-
-    def get_function(self):
-        def week(vals):
-            warnings.filterwarnings("ignore",
-                                    message=("Series.dt.weekofyear and Series.dt.week "
-                                             "have been deprecated."),
-                                    module="featuretools"
-                                    )
-            return vals.dt.week
-        return week
-
-
-class Month(TransformPrimitive):
-    """Determines the month value of a datetime.
-
-    Examples:
-        >>> from datetime import datetime
-        >>> dates = [datetime(2019, 3, 1),
-        ...          datetime(2019, 6, 17, 11, 10, 50),
-        ...          datetime(2019, 11, 30, 19, 45, 15)]
-        >>> month = Month()
-        >>> month(dates).tolist()
-        [3, 6, 11]
-    """
-    name = "month"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(logical_type=Ordinal(order=list(range(1, 13))), semantic_tags={'category'})
-    compatibility = [Library.PANDAS, Library.DASK, Library.KOALAS]
-    description_template = "the month of {}"
-
-    def get_function(self):
-        def month(vals):
-            return vals.dt.month
-        return month
-
-
-class Year(TransformPrimitive):
-    """Determines the year value of a datetime.
-
-    Examples:
-        >>> from datetime import datetime
-        >>> dates = [datetime(2019, 3, 1),
-        ...          datetime(2048, 6, 17, 11, 10, 50),
-        ...          datetime(1950, 11, 30, 19, 45, 15)]
-        >>> year = Year()
-        >>> year(dates).tolist()
-        [2019, 2048, 1950]
-    """
-    name = "year"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(logical_type=Ordinal(order=list(range(1, 3000))), semantic_tags={'category'})
-    compatibility = [Library.PANDAS, Library.DASK, Library.KOALAS]
-    description_template = "the year of {}"
-
-    def get_function(self):
-        def year(vals):
-            return vals.dt.year
-        return year
-
-
-class IsWeekend(TransformPrimitive):
-    """Determines if a date falls on a weekend.
-
-    Examples:
-        >>> from datetime import datetime
-        >>> dates = [datetime(2019, 3, 1),
-        ...          datetime(2019, 6, 17, 11, 10, 50),
-        ...          datetime(2019, 11, 30, 19, 45, 15)]
-        >>> is_weekend = IsWeekend()
-        >>> is_weekend(dates).tolist()
-        [False, False, True]
-    """
-    name = "is_weekend"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(logical_type=BooleanNullable)
-    compatibility = [Library.PANDAS, Library.DASK, Library.KOALAS]
-    description_template = "whether {} occurred on a weekend"
-
-    def get_function(self):
-        def is_weekend(vals):
-            return vals.dt.weekday > 4
-        return is_weekend
-
-
-class Weekday(TransformPrimitive):
-    """Determines the day of the week from a datetime.
-
-    Description:
-        Returns the day of the week from a datetime value. Weeks
-        start on Monday (day 0) and run through Sunday (day 6).
-
-    Examples:
-        >>> from datetime import datetime
-        >>> dates = [datetime(2019, 3, 1),
-        ...          datetime(2019, 6, 17, 11, 10, 50),
-        ...          datetime(2019, 11, 30, 19, 45, 15)]
-        >>> weekday = Weekday()
-        >>> weekday(dates).tolist()
-        [4, 0, 5]
-    """
-    name = "weekday"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(logical_type=Ordinal(order=list(range(7))), semantic_tags={'category'})
-    compatibility = [Library.PANDAS, Library.DASK, Library.KOALAS]
-    description_template = "the day of the week of {}"
-
-    def get_function(self):
-        def weekday(vals):
-            return vals.dt.weekday
-        return weekday
 
 
 class NumCharacters(TransformPrimitive):
@@ -377,53 +96,6 @@ class NumWords(TransformPrimitive):
         def word_counter(array):
             return array.fillna('').str.count(' ') + 1
         return word_counter
-
-
-class TimeSince(TransformPrimitive):
-    """Calculates time from a value to a specified cutoff datetime.
-
-    Args:
-        unit (str): Defines the unit of time to count from.
-            Defaults to Seconds. Acceptable values:
-            years, months, days, hours, minutes, seconds, milliseconds, nanoseconds
-
-    Examples:
-        >>> from datetime import datetime
-        >>> time_since = TimeSince()
-        >>> times = [datetime(2019, 3, 1, 0, 0, 0, 1),
-        ...          datetime(2019, 3, 1, 0, 0, 1, 0),
-        ...          datetime(2019, 3, 1, 0, 2, 0, 0)]
-        >>> cutoff_time = datetime(2019, 3, 1, 0, 0, 0, 0)
-        >>> values = time_since(times, time=cutoff_time)
-        >>> list(map(int, values))
-        [0, -1, -120]
-
-        Change output to nanoseconds
-
-        >>> from datetime import datetime
-        >>> time_since_nano = TimeSince(unit='nanoseconds')
-        >>> times = [datetime(2019, 3, 1, 0, 0, 0, 1),
-        ...          datetime(2019, 3, 1, 0, 0, 1, 0),
-        ...          datetime(2019, 3, 1, 0, 2, 0, 0)]
-        >>> cutoff_time = datetime(2019, 3, 1, 0, 0, 0, 0)
-        >>> values = time_since_nano(times, time=cutoff_time)
-        >>> list(map(lambda x: int(round(x)), values))
-        [-1000, -1000000000, -120000000000]
-    """
-    name = 'time_since'
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(semantic_tags={'numeric'})
-    uses_calc_time = True
-    compatibility = [Library.PANDAS, Library.DASK]
-    description_template = "the time from {} to the cutoff time"
-
-    def __init__(self, unit="seconds"):
-        self.unit = unit.lower()
-
-    def get_function(self):
-        def pd_time_since(array, time):
-            return convert_time_units((time - array).dt.total_seconds(), self.unit)
-        return pd_time_since
 
 
 class IsIn(TransformPrimitive):
@@ -551,91 +223,6 @@ class Percentile(TransformPrimitive):
 
     def get_function(self):
         return lambda array: array.rank(pct=True)
-
-
-class Haversine(TransformPrimitive):
-    """Calculates the approximate haversine distance between two LatLong columns.
-
-        Args:
-            unit (str): Determines the unit value to output. Could
-                be `miles` or `kilometers`. Default is `miles`.
-
-        Examples:
-            >>> haversine = Haversine()
-            >>> distances = haversine([(42.4, -71.1), (40.0, -122.4)],
-            ...                       [(40.0, -122.4), (41.2, -96.75)])
-            >>> np.round(distances, 3).tolist()
-            [2631.231, 1343.289]
-
-            Output units can be specified
-
-            >>> haversine_km = Haversine(unit='kilometers')
-            >>> distances_km = haversine_km([(42.4, -71.1), (40.0, -122.4)],
-            ...                             [(40.0, -122.4), (41.2, -96.75)])
-            >>> np.round(distances_km, 3).tolist()
-            [4234.555, 2161.814]
-    """
-    name = 'haversine'
-    input_types = [ColumnSchema(logical_type=LatLong), ColumnSchema(logical_type=LatLong)]
-    return_type = ColumnSchema(semantic_tags={'numeric'})
-    commutative = True
-
-    def __init__(self, unit='miles'):
-        valid_units = ['miles', 'kilometers']
-        if unit not in valid_units:
-            error_message = 'Invalid unit %s provided. Must be one of %s' % (unit, valid_units)
-            raise ValueError(error_message)
-        self.unit = unit
-        self.description_template = "the haversine distance in {} between {{}} and {{}}".format(self.unit)
-
-    def get_function(self):
-        def haversine(latlong_1, latlong_2):
-            lat_1s, lon_1s = _deconstrct_latlongs(latlong_1)
-            lat_2s, lon_2s = _deconstrct_latlongs(latlong_2)
-            distance = _haversine_calculate(lat_1s, lon_1s, lat_2s, lon_2s, self.unit)
-            return distance
-        return haversine
-
-    def generate_name(self, base_feature_names):
-        name = u"{}(".format(self.name.upper())
-        name += u", ".join(base_feature_names)
-        if self.unit != 'miles':
-            name += u", unit={}".format(self.unit)
-        name += u")"
-        return name
-
-
-class Age(TransformPrimitive):
-    """Calculates the age in years as a floating point number given a
-       date of birth.
-
-    Description:
-        Age in years is computed by calculating the number of days between
-        the date of birth and the reference time and dividing the result
-        by 365.
-
-    Examples:
-        Determine the age of three people as of Jan 1, 2019
-        >>> import pandas as pd
-        >>> reference_date = pd.to_datetime("01-01-2019")
-        >>> age = Age()
-        >>> input_ages = [pd.to_datetime("01-01-2000"),
-        ...               pd.to_datetime("05-30-1983"),
-        ...               pd.to_datetime("10-17-1997")]
-        >>> age(input_ages, time=reference_date).tolist()
-        [19.013698630136986, 35.61643835616438, 21.221917808219178]
-    """
-    name = "age"
-    input_types = [ColumnSchema(logical_type=Datetime, semantic_tags={'date_of_birth'})]
-    return_type = ColumnSchema(logical_type=AgeFractional, semantic_tags={'numeric'})
-    uses_calc_time = True
-    compatibility = [Library.PANDAS, Library.DASK]
-    description_template = "the age from {}"
-
-    def get_function(self):
-        def age(x, time=None):
-            return (time - x).dt.days / 365
-        return age
 
 
 class URLToDomain(TransformPrimitive):
@@ -855,247 +442,3 @@ class NumericLag(TransformPrimitive):
             x = pd.Series(numeric.values, index=time_index.values)
             return x.shift(periods=self.periods, fill_value=self.fill_value).values
         return lag
-
-
-class IsInGeoBox(TransformPrimitive):
-    """Determines if coordinates are inside a box defined by two
-    corner coordinate points.
-
-    Description:
-        Coordinate values should be specified as (latitude, longitude)
-        tuples. This primitive is unable to handle coordinates and boxes
-        at the poles, and near +/- 180 degrees latitude.
-
-    Args:
-        point1 (tuple(float, float)): The coordinates
-            of the first corner of the box. Defaults to (0, 0).
-        point2 (tuple(float, float)): The coordinates
-            of the diagonal corner of the box. Defaults to (0, 0).
-
-    Example:
-        >>> is_in_geobox = IsInGeoBox((40.7128, -74.0060), (42.2436, -71.1677))
-        >>> is_in_geobox([(41.034, -72.254), (39.125, -87.345)]).tolist()
-        [True, False]
-    """
-    name = "is_in_geobox"
-    input_types = [ColumnSchema(logical_type=LatLong)]
-    return_type = ColumnSchema(logical_type=BooleanNullable)
-
-    def __init__(self, point1=(0, 0), point2=(0, 0)):
-        self.point1 = point1
-        self.point2 = point2
-        self.lats = np.sort(np.array([point1[0], point2[0]]))
-        self.lons = np.sort(np.array([point1[1], point2[1]]))
-
-    def get_function(self):
-        def geobox(latlongs):
-            if latlongs.hasnans:
-                latlongs = np.where(latlongs.isnull(), pd.Series([(np.nan, np.nan)] * len(latlongs)), latlongs)
-            transposed = np.transpose([list(latlon) for latlon in latlongs])
-            lats = (self.lats[0] <= transposed[0]) & \
-                   (self.lats[1] >= transposed[0])
-            longs = (self.lons[0] <= transposed[1]) & \
-                    (self.lons[1] >= transposed[1])
-            return lats & longs
-        return geobox
-
-
-class GeoMidpoint(TransformPrimitive):
-    """Determines the geographic center of two coordinates.
-
-    Examples:
-        >>> geomidpoint = GeoMidpoint()
-        >>> geomidpoint([(42.4, -71.1)], [(40.0, -122.4)])
-        [(41.2, -96.75)]
-    """
-    name = "geomidpoint"
-    input_types = [ColumnSchema(logical_type=LatLong), ColumnSchema(logical_type=LatLong)]
-    return_type = ColumnSchema(logical_type=LatLong)
-    commutative = True
-
-    def get_function(self):
-        def geomidpoint_func(latlong_1, latlong_2):
-            lat_1s, lon_1s = _deconstrct_latlongs(latlong_1)
-            lat_2s, lon_2s = _deconstrct_latlongs(latlong_2)
-            lat_middle = np.array([lat_1s, lat_2s]).transpose().mean(axis=1)
-            lon_middle = np.array([lon_1s, lon_2s]).transpose().mean(axis=1)
-            return list(zip(lat_middle, lon_middle))
-        return geomidpoint_func
-
-
-class CityblockDistance(TransformPrimitive):
-    """Calculates the distance between points in a city road grid.
-
-    Description:
-        This distance is calculated using the haversine formula, which
-        takes into account the curvature of the Earth.
-        If either input data contains `NaN`s, the calculated
-        distance with be `NaN`.
-        This calculation is also known as the Mahnattan distance.
-
-    Args:
-        unit (str): Determines the unit value to output. Could
-            be miles or kilometers. Default is miles.
-
-    Examples:
-        >>> cityblock_distance = CityblockDistance()
-        >>> DC = (38, -77)
-        >>> Boston = (43, -71)
-        >>> NYC = (40, -74)
-        >>> distances_mi = cityblock_distance([DC, DC], [NYC, Boston])
-        >>> np.round(distances_mi, 3).tolist()
-        [301.519, 672.089]
-
-        We can also change the units in which the distance is calculated.
-
-        >>> cityblock_distance_kilometers = CityblockDistance(unit='kilometers')
-        >>> distances_km = cityblock_distance_kilometers([DC, DC], [NYC, Boston])
-        >>> np.round(distances_km, 3).tolist()
-        [485.248, 1081.622]
-    """
-    name = "cityblock_distance"
-    input_types = [ColumnSchema(logical_type=LatLong), ColumnSchema(logical_type=LatLong)]
-    return_type = ColumnSchema(logical_type=Double, semantic_tags={'numeric'})
-    commutative = True
-
-    def __init__(self, unit='miles'):
-        if unit not in ['miles', 'kilometers']:
-            raise ValueError("Invalid unit given")
-        self.unit = unit
-
-    def get_function(self):
-        def cityblock(latlong_1, latlong_2):
-            lat_1s, lon_1s = _deconstrct_latlongs(latlong_1)
-            lat_2s, lon_2s = _deconstrct_latlongs(latlong_2)
-            lon_dis = _haversine_calculate(lat_1s, lon_1s, lat_1s, lon_2s,
-                                           self.unit)
-            lat_dist = _haversine_calculate(lat_1s, lon_1s, lat_2s, lon_1s,
-                                            self.unit)
-            return pd.Series(lon_dis + lat_dist)
-        return cityblock
-
-
-class DateToHoliday(TransformPrimitive):
-    """Transforms time of an instance into the holiday name, if there is one.
-
-    Description:
-        If there is no holiday, it returns `NaN`. Currently only works for the
-        United States and Canada with dates between 1950 and 2100.
-
-    Args:
-        country (str): Country to use for determining Holidays.
-            Default is 'US'. Should be one of the available countries here:
-            https://github.com/dr-prodigy/python-holidays#available-countries
-
-    Examples:
-        >>> from datetime import datetime
-        >>> date_to_holiday = DateToHoliday()
-        >>> dates = pd.Series([datetime(2016, 1, 1),
-        ...          datetime(2016, 2, 27),
-        ...          datetime(2017, 5, 29, 10, 30, 5),
-        ...          datetime(2018, 7, 4)])
-        >>> date_to_holiday(dates).tolist()
-        ["New Year's Day", nan, 'Memorial Day', 'Independence Day']
-
-        We can also change the country.
-
-        >>> date_to_holiday_canada = DateToHoliday(country='Canada')
-        >>> dates = pd.Series([datetime(2016, 7, 1),
-        ...          datetime(2016, 11, 15),
-        ...          datetime(2017, 12, 26),
-        ...          datetime(2018, 9, 3)])
-        >>> date_to_holiday_canada(dates).tolist()
-        ['Canada Day', nan, 'Boxing Day', 'Labour Day']
-    """
-    name = "date_to_holiday"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(logical_type=Categorical, semantic_tags={'category'})
-
-    def __init__(self, country='US'):
-        self.country = country
-        self.holidayUtil = HolidayUtil(country)
-
-    def get_function(self):
-        def date_to_holiday(x):
-            holiday_df = self.holidayUtil.to_df()
-            df = pd.DataFrame({'date': x})
-            df.date = df.date.dt.normalize().astype('datetime64')
-
-            df = df.merge(holiday_df, how='left', left_on='date', right_on='holiday_date')
-            return df.names.values
-        return date_to_holiday
-
-
-class DistanceToHoliday(TransformPrimitive):
-    """Computes the number of days before or after a given holiday.
-
-    Description:
-        For a list of dates, return the distance from the nearest
-        occurrence of a chosen holiday. The distance is returned in
-        days. If the closest occurrence is prior to the date given,
-        return a negative number.
-
-        If a date is missing, return `NaN`.
-
-        Currently only works with dates between 1950 and 2100.
-
-    Args:
-        holiday (str): Name of the holiday. Defaults to New Year's Day.
-
-        country (str): Specifies which country's calendar to use for the
-            given holiday. Default is `US`.
-
-    Examples:
-        >>> from datetime import datetime
-        >>> distance_to_holiday = DistanceToHoliday("New Year's Day")
-        >>> dates = [datetime(2010, 1, 1),
-        ...          datetime(2012, 5, 31),
-        ...          datetime(2017, 7, 31),
-        ...          datetime(2020, 12, 31)]
-        >>> distance_to_holiday(dates).tolist()
-        [0, -151, 154, 1]
-
-        We can also control the country in which we're searching for
-            a holiday.
-
-        >>> distance_to_holiday = DistanceToHoliday("Victoria Day", country='Canada')
-        >>> dates = [datetime(2010, 1, 1),
-        ...          datetime(2012, 5, 31),
-        ...          datetime(2017, 7, 31),
-        ...          datetime(2020, 12, 31)]
-        >>> distance_to_holiday(dates).tolist()
-        [143, -10, -70, 144]
-    """
-    name = "distance_to_holiday"
-    input_types = [ColumnSchema(logical_type=Datetime)]
-    return_type = ColumnSchema(semantic_tags={'numeric'})
-    default_value = 0
-
-    def __init__(self, holiday="New Year's Day", country="US"):
-        self.country = country
-        self.holiday = holiday
-        self.holidayUtil = HolidayUtil(country)
-
-        available_holidays = list(set(self.holidayUtil.federal_holidays.values()))
-        if self.holiday not in available_holidays:
-            error = 'must be one of the available holidays:\n%s' % available_holidays
-            raise ValueError(error)
-
-    def get_function(self):
-        def distance_to_holiday(x):
-            holiday_df = self.holidayUtil.to_df()
-            holiday_df = holiday_df[holiday_df.names == self.holiday]
-
-            df = pd.DataFrame({'date': x})
-            df['x_index'] = df.index  # store original index as a column
-            df = df.dropna()
-            df = df.sort_values('date')
-            df.date = df.date.dt.normalize()
-
-            matches = pd.merge_asof(df, holiday_df, left_on='date', right_on='holiday_date',
-                                    direction='nearest', tolerance=pd.Timedelta('365d'))
-            matches = matches.set_index('x_index')
-            matches['days_diff'] = (matches.holiday_date - matches.date).dt.days
-
-            return matches.days_diff.reindex_like(x)
-        return distance_to_holiday
